requirements.txt
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
fonttools==4.53.1
idna==3.8
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
matplotlib==3.9.2
msgpack==1.0.8
noisereduce==3.0.2
numba==0.60.0
numpy==2.0.2
packaging==24.1
pillow==10.4.0
platformdirs==4.3.2
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
